Extra factory class for V3 AWS SDK.
